requirements.txt
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
pyarrow==15.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
black==24.1.1
flake8==7.0.0
